replace old cap bank
