+++ enderio/src/datagen/java/com/enderio/enderio/datagen/client/models/EIOBlockStateProvider.java
@@ -210,10 +210,6 @@ public class EIOBlockStateProvider extends BlockStateProvider {
             simpleBlockWithExistingModel(capacitorBank.get());
         }
 
-        for (var capacitorBank : EIOBlocks.NEW_CAPACITOR_BANKS.values()) {
-            simpleBlockWithExistingModel(capacitorBank.get());
-        }
-
         // Niard
         machineBlock(EIOBlocks.NIARD.get());
