+++ enderio/src/main/java/com/enderio/enderio/client/content/machines/gui/widget/EnergyWidget.java
@@ -3,7 +3,6 @@ package com.enderio.enderio.client.content.machines.gui.widget;
 import com.enderio.core.client.gui.widgets.EIOWidget;
 import com.enderio.core.common.util.TooltipUtil;
 import com.enderio.enderio.EnderIO;
-import com.enderio.enderio.content.machines.capacitor_bank.rework.NewCapacitorBankMenu;
 import com.enderio.enderio.foundation.io.energy.ILargeMachineEnergyStorage;
 import com.enderio.enderio.foundation.lang.EIOCommonLang;
 import com.mojang.blaze3d.systems.RenderSystem;
