+++ enderio/src/main/java/com/enderio/enderio/content/machines/capacitor_bank/CapacitorBankMenu.java
@@ -1,26 +1,114 @@
 package com.enderio.enderio.content.machines.capacitor_bank;
 
-import com.enderio.enderio.foundation.menu.legacy.LegacyPoweredMachineMenu;
+import com.enderio.core.common.menu.BaseBlockEntityMenu;
+import com.enderio.core.common.network.menu.EnumSyncSlot;
+import com.enderio.enderio.api.io.RedstoneControl;
+import com.enderio.enderio.foundation.io.energy.ILargeMachineEnergyStorage;
+import com.enderio.enderio.init.EIOBlockEntities;
 import com.enderio.enderio.init.EIOMenus;
-import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.level.block.entity.BlockEntity;
-import org.apache.logging.log4j.LogManager;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.neoforged.neoforge.energy.IEnergyStorage;
+import net.neoforged.neoforge.registries.DeferredHolder;
 
-public class CapacitorBankMenu extends LegacyPoweredMachineMenu<CapacitorBankBlockEntity> {
-    public CapacitorBankMenu(int containerId, @Nullable CapacitorBankBlockEntity blockEntity, Inventory inventory) {
-        super(EIOMenus.CAPACITOR_BANK.get(), containerId, blockEntity, inventory);
+import java.util.Objects;
+
+public class CapacitorBankMenu extends BaseBlockEntityMenu<CapacitorBankBlockEntity> {
+
+    private final EnumSyncSlot<RedstoneControl> redstoneControlSlot;
+
+    public CapacitorBankMenu(int containerId, Inventory playerInventory, CapacitorBankBlockEntity blockEntity) {
+        super(EIOMenus.CAPACITOR_BANK.get(), containerId, playerInventory, blockEntity);
+
+        redstoneControlSlot = addUpdatableSyncSlot(EnumSyncSlot.simple(RedstoneControl.class,
+            blockEntity::getRedstoneControl, blockEntity::setNetworkRedstoneControl));
+
         addPlayerInventorySlots(8, 84);
     }
 
-    public static CapacitorBankMenu factory(int containerId, Inventory inventory, FriendlyByteBuf buf) {
-        BlockEntity entity = inventory.player.level().getBlockEntity(buf.readBlockPos());
-        if (entity instanceof CapacitorBankBlockEntity castBlockEntity) {
-            return new CapacitorBankMenu(containerId, castBlockEntity, inventory);
+    public CapacitorBankMenu(int containerId, Inventory playerInventory, RegistryFriendlyByteBuf buf) {
+        super(EIOMenus.CAPACITOR_BANK.get(), containerId, playerInventory, buf,
+            EIOBlockEntities.CAPACITOR_BANKS.values().stream().map(DeferredHolder::get).toArray(BlockEntityType[]::new));
+
+        redstoneControlSlot = addUpdatableSyncSlot(EnumSyncSlot.standalone(RedstoneControl.class));
+
+        addPlayerInventorySlots(8, 84);
+    }
+
+    @Override
+    public boolean stillValid(Player player) {
+        return Container.stillValidBlockEntity(getBlockEntity(), player);
+    }
+
+    //No inv
+    @Override
+    public ItemStack quickMoveStack(Player player, int index) {
+        return ItemStack.EMPTY;
+    }
+
+    public EnergyStorage getEnergyStorage() {
+        CapacitorBankManager.CapacitorSyncData data = CapacitorBankManager.getData(getBlockEntity().getUuid());
+        if (data == null) {
+            return EnergyStorage.EMPTY;
+        }
+        return new EnergyStorage(data.storedEnergy(), data.capacity());
+    }
+
+    public RedstoneControl getRedstoneControl() {
+        return Objects.requireNonNull(redstoneControlSlot).get();
+    }
+
+    public void setRedstoneControl(RedstoneControl redstoneControl) {
+        Objects.requireNonNull(redstoneControlSlot).set(redstoneControl);
+        updateSlot(redstoneControlSlot);
+    }
+
+    public record EnergyStorage(Long energyStored, Long maxEnergyStored) implements IEnergyStorage, ILargeMachineEnergyStorage {
+
+        public static final EnergyStorage EMPTY = new EnergyStorage(0L,0L);
+
+        @Override
+        public int receiveEnergy(int i, boolean b) {
+            return 0;
         }
 
-        LogManager.getLogger().warn("couldn't find BlockEntity");
-        return new CapacitorBankMenu(containerId, null, inventory);
+        @Override
+        public int extractEnergy(int i, boolean b) {
+            return 0;
+        }
+
+        @Override
+        public int getEnergyStored() {
+            return Math.clamp(energyStored, 0, Integer.MAX_VALUE);
+        }
+
+        @Override
+        public int getMaxEnergyStored() {
+            return Math.clamp(maxEnergyStored, 0, Integer.MAX_VALUE);
+        }
+
+        @Override
+        public boolean canExtract() {
+            return false;
+        }
+
+        @Override
+        public boolean canReceive() {
+            return false;
+        }
+
+        @Override
+        public long getLargeEnergyStored() {
+            return energyStored;
+        }
+
+        @Override
+        public long getLargeMaxEnergyStored() {
+            return maxEnergyStored;
+        }
     }
 }
